--- src/main/java/com/linag/ciclo3/modelos/Empleado.java
@@ -1,75 +1,0 @@
-package com.linag.ciclo3.modelos;
-
-
-import javax.persistence.*;
-
-@Entity
-@Table(name = "Empleado")
-public class Empleado {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY )
-    //@GeneratedValue(strategy = GenerationType.AUTO )
-
-    private int id;
-    private String nombre;
-    private String correo;
-    @ManyToOne
-    @JoinColumn(name = "empresa_id")
-    private Empresa empresa;
-    private String rol;
-
-    //Costructor vacio
-
-
-    public Empleado() {
-    }
-
-    public Empleado(String nombre, String correo, Empresa empresa, String rol) {
-        this.nombre = nombre;
-        this.correo = correo;
-        this.empresa = empresa;
-        this.rol = rol;
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getNombre() {
-        return nombre;
-    }
-
-    public void setNombre(String nombre) {
-        this.nombre = nombre;
-    }
-
-    public String getCorreo() {
-        return correo;
-    }
-
-    public void setCorreo(String correo) {
-        this.correo = correo;
-    }
-
-    public String getRol() {
-        return rol;
-    }
-
-    public void setRol(String rol) {
-        this.rol = rol;
-    }
-
-    public Empresa getEmpresa() {
-        return empresa;
-    }
-
-    public void setEmpresa(Empresa empresa) {
-        this.empresa = empresa;
-    }
-
-}
