+++ src/main/java/com/linag/ciclo3/repositories/EmpleadoRepository.java
@@ -1,5 +1,6 @@
 package com.linag.ciclo3.repositories;
 
+/*
 import com.linag.ciclo3.entities.Empleado;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
@@ -8,3 +9,4 @@ import org.springframework.stereotype.Repository;
 
 public interface EmpleadoRepository extends CrudRepository<Empleado,Integer> {
 }
+*/
