+++ src/main/java/com/linag/ciclo3/repositories/MovimientosRepository.java
@@ -1,5 +1,6 @@
 package com.linag.ciclo3.repositories;
 
+/*
 import com.linag.ciclo3.entities.MovimientoDinero;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -8,3 +9,4 @@ import org.springframework.stereotype.Repository;
 
 public interface MovimientosRepository  extends JpaRepository<MovimientoDinero,Integer> {
 }
+*/
