+++ src/main/java/com/linag/ciclo3/services/EmpleadoService.java
@@ -1,7 +1,9 @@
-package com.linag.ciclo3.service;
-
+package com.linag.ciclo3.services;
+/*
+import com.linag.ciclo3.entities.Empleado;
 import com.linag.ciclo3.modelos.Empleado;
 import com.linag.ciclo3.repo.EmpleadoRepository;
+import com.linag.ciclo3.repositories.EmpleadoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -45,4 +47,4 @@ public class EmpleadoService {
         }
         return true;
     }
-}
+}*/
