+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.datasource.url=jdbc:postgresql://localhost:5432/clase3
+spring.datasource.url=jdbc:postgresql://localhost:5432/ciclo3
 spring.datasource.username=postgres
 spring.datasource.password=li22173$
 spring.jpa.hibernate.ddl-auto=update
